Minor refactoring

+++ medical_image_converter/converter.py
@@ -17,13 +17,17 @@ def convert_dicom_to_jpg(input_path:str, output_path:str) ->str:
             input_path: input path of the .ima image
             output_path: output path where the image is stored
     """
+    if not os.path.exists(input_path):
+        raise FileNotFoundError("File not found")
+
     try:
+        print("Converting {} to {}".format(input_path, output_path))
         ds = dicom.read_file(input_path,)
         pixel_array_numpy = ds.pixel_array
         cv2.imwrite(output_path, pixel_array_numpy)
         return output_path
     except Exception as e:
-        print("ERROR:", str(e))
+        print("Error while converting dicom to jpg.:", str(e))
         return ''
 
 def multithread_converter(source_list:list, dest_list:list)->None:
